pyproject.toml
[project]
name = "krono"
version = "0.1.0"
description = ""
authors = [
    {name = "pongf456",email = "adonayyeidersonparedes@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-telegram-bot (>=22.5,<23.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "pytz (>=2025.2,<2026.0)"
]

[tool.poetry]
packages = [{include = "krono", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
